Move start background server to a common folder
