+++ .dockerignore
@@ -1,3 +1,4 @@
 node_modules/*
 public/bower_components/*
 build/*
+.git
