+++ common/backgroundServer.js
