+++ oauth-proxy/tests/e2e.test.js
@@ -5,7 +5,7 @@ const request = require('request-promise-native');
 const { Issuer } = require('openid-client');
 const { randomBytes } = require('crypto');
 
-const { buildBackgroundServerModule } = require('./backgroundServer');
+const { buildBackgroundServerModule } = require('../../common/backgroundServer');
 const upstreamOAuthTestServer = require('./upstreamOAuthTestServer');
 const { startServerInBackground, stopBackgroundServer } = buildBackgroundServerModule("oauth-proxy test app");
 const { buildApp } = require('../index');
