+++ saml-proxy/Dockerfile
@@ -5,14 +5,15 @@ WORKDIR /opt/app/
 
 RUN apk add git python make g++
 RUN git config --global url."https://".insteadOf git://
-ADD ./package.json package.json
-ADD ./package-lock.json package-lock.json
+ADD ./saml-proxy/package.json package.json
+ADD ./saml-proxy/package-lock.json package-lock.json
 RUN npm install
 
-ADD ./certs/* /usr/local/share/ca-certificates/
+ADD ./saml-proxy/certs/* /usr/local/share/ca-certificates/
 RUN update-ca-certificates
 
-ADD ./ .
+ADD ./saml-proxy ./
+ADD ./common ../common
 
 RUN cat certs/*.pem > certs/combined.pem
 ENV NODE_EXTRA_CA_CERTS /opt/app/certs/combined.pem
